refactor: 페이지 mutations entity로 이동
