+++ apps/frontend/src/entities/page/index.ts
@@ -5,6 +5,7 @@ export {
   deletePage,
   updatePage,
 } from "./api/pageApi";
+export { useCreatePage, useDeletePage } from "./model/pageMutations";
 
 export { usePageStore } from "./model/pageStore";
 export { type Page, type CreatePageRequest } from "./model/pageTypes";
