+++ apps/frontend/src/features/pageSidebar/index.ts
@@ -1,5 +1,3 @@
-export { useCreatePage } from "./api/usePages";
-
 export { LogoBtn } from "./ui/LogoBtn";
 export { NoteList } from "./ui/NoteList";
 export { Tools } from "./ui/Tools";
